moved logic to file-utils
