+++ site/app/libraries/FileUtils.php
@@ -13,7 +13,7 @@ class FileUtils {
     const IGNORE_FOLDERS = [".svn", ".git", ".idea", "__macosx"];
     const IGNORE_FILES = ['.ds_store'];
     const ALLOWED_IMAGE_TYPES = ['jpg', 'jpeg', 'png', 'gif'];
-    // also update in 'ta-grading-panels-init.ts'
+    // also update in 'file-utils.ts'
     const SUBMISSION_META_FILES = ['.submit.notebook', '.submit.timestamp', '.submit.VCS_CHECKOUT',
     '.user_assignment_access.json', '.bulk_upload_data.json'];
 
@@ -451,7 +451,7 @@ class FileUtils {
         return preg_replace('#' . preg_quote($sep) . '+#', $sep, join($sep, $paths));
     }
 
-    // also update in 'ta-grading-panels-init.ts'
+    // also update in 'file-utils.ts'
     /**
      * Given a filename, checks whether it is a Submitty metadata file.
      * @param string $filename
